backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
email-validator>=2.1.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
